moved test_cli_entry to test_upload, and rewrote upload tests
